requirements.txt
torch>=1.9.0
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0
pyyaml>=5.4.0
tensorboard>=2.5.0
tqdm>=4.60.0